requirements.txt
ipykernel
ipywidgets
pandas
seaborn
matplotlib
Pillow
tensorflow < 3
opencv-python
black
pydot
tqdm
scikit-learn


